% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateTuningTime.R
\name{estimateTuneRangerTime}
\alias{estimateTuneRangerTime}
\title{estimateTuneRangerTime}
\usage{
estimateTuneRangerTime(formula, data, iters = 100, num.threads = 1,
  num.trees = 1000, respect.unordered.factors = TRUE)
}
\arguments{
\item{formula}{}

\item{data}{}

\item{iters}{}

\item{num.threads}{}

\item{num.trees}{}

\item{respect.unordered.factors}{}
}
\value{
estimated time for the tuning procedure
}
\description{
estimateTuneRangerTime
}
