% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateTuneRFTime.R
\name{estimateTuneRFTime}
\alias{estimateTuneRFTime}
\title{estimateTuneRFTime}
\usage{
estimateTuneRFTime(task, iters = 100, num.threads = 1, num.trees = 1000,
  respect.unordered.factors = TRUE)
}
\arguments{
\item{task}{The mlr task created by makeClassifTask or makeRegrTask.}

\item{iters}{Number of iterations.}

\item{num.threads}{Number of threads. Default is 1.}

\item{num.trees}{Number of trees.}

\item{respect.unordered.factors}{Handling of unordered factor covariates. One of 'ignore', 'order' and 'partition'. For the "extratrees" splitrule the default is "partition" for all other splitrules 'ignore'. Alternatively TRUE (='order') or FALSE (='ignore') can be used. See below for details.}
}
\value{
estimated time for the tuning procedure
}
\description{
estimateTuneRFTime
}
\examples{
estimateTuneRFTime(iris.task)
}
