% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneRanger.R
\name{tuneRanger}
\alias{tuneRanger}
\title{tuneRanger}
\usage{
tuneRanger(formula, data, measure = measureMSE, num.threads = 1,
  num.trees = 1000, replace = TRUE, iters = 100,
  save.file.path = "./optpath.RData")
}
\arguments{
\item{formula}{}

\item{data}{}

\item{measure}{}

\item{num.threads}{}

\item{num.trees}{}

\item{replace}{}

\item{iters}{}

\item{save.file.path}{}
}
\value{
res
}
\description{
tuneRanger
}
