% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneRF.R
\name{tuneRF}
\alias{tuneRF}
\title{tuneRF}
\usage{
tuneRF(task, measure = NULL, iters = 100, num.trees = 1000,
  num.threads = 1, replace = TRUE, save.file.path = "./optpath.RData")
}
\arguments{
\item{task}{}

\item{measure}{}

\item{iters}{}

\item{num.trees}{}

\item{num.threads}{}

\item{replace}{}

\item{save.file.path}{}
}
\value{
res
}
\description{
tuneRF
}
