% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateTimeTuneRanger.R
\name{estimateTimeTuneRanger}
\alias{estimateTimeTuneRanger}
\title{estimateTimeTuneRanger}
\usage{
estimateTimeTuneRanger(task, iters = 100, num.threads = 1,
  num.trees = 1000, respect.unordered.factors = FALSE)
}
\arguments{
\item{task}{The mlr task created by makeClassifTask or makeRegrTask.}

\item{iters}{Number of iterations.}

\item{num.threads}{Number of threads. Default is 1.}

\item{num.trees}{Number of trees.}

\item{respect.unordered.factors}{Handling of unordered factor covariates. One of 'ignore', 'order' and 'partition'. For the "extratrees" splitrule the default is "partition" for all other splitrules 'ignore'. Alternatively TRUE (='order') or FALSE (='ignore') can be used. See below for details.}
}
\value{
estimated time for the tuning procedure
}
\description{
estimateTimeTuneRanger
}
\examples{
estimateTimeTuneRanger(iris.task)
}
